favorite stuff added ver 2

+++ app/src/main/java/com/example/asus/converterapp/FavoriteList.java
@@ -3,20 +3,20 @@ package com.example.asus.converterapp;
 public class FavoriteList {
     private String spinnerFrom;
     private String spinnerTo;
-    private int input;
-    private int result;
+    private double input;
+    private double result;
 
-    public FavoriteList(String spinnerFrom,String spinnerTo, int input, int result ){
+    public FavoriteList(String spinnerFrom, String spinnerTo, double input, double result ){
         this.spinnerFrom = spinnerFrom;
         this.spinnerTo = spinnerTo;
         this.input = input;
         this.result = result;
     }
     public String getInput(){
-        return Integer.toString(this.input);
+        return Double.toString(this.input);
     }
     public String getResult(){
-        return Integer.toString(this.result);
+        return Double.toString(this.result);
     }
     public String getSpinnerFrom(){
         return this.spinnerFrom;
